requirements.txt
Django>=2.1.3,<=3.4.0
djangorestframework>=3.9.0,<=3.11.0
psycopg2>=2.7.5,<2.8.0
django-robots==3.1.0
six==1.13.0

flake8>=3.6.0,<=3.7.0
